+++ src/Services/MarketServices.php
@@ -2,13 +2,15 @@
 
 namespace App\Services;
 
-use App\Traits\ConsumesExternalService;
+use App\Traits\AuthorizesMarketRequests, App\Traits\ConsumesExternalService, App\Traits\InteractsWithMarketResponses;
 use Exception;
 use stdClass;
 
 class MarketServices
 {
     use ConsumesExternalService;
+    use InteractsWithMarketResponses;
+    use AuthorizesMarketRequests;
 
     protected string $baseUri;
     private $passwordClientSecret;
@@ -32,46 +34,5 @@ class MarketServices
         $this->passwordClientId = $passwordClientId;
     }
 
-    /**
-     * @param array $responde
-     * @return void
-     * @throws Exception
-     */
-    public function CheckIfErrorResponse(array $responde): void
-    {
-        if(isset($responde->error)){
-            throw new Exception("Fallo en la respuesta {$responde->error}" );
-        }
-    }
 
-    /**
-     * @param string $response
-     * @return stdClass
-     */
-    public function decodeResponse(string $response): stdClass
-    {
-        $decodeResponse = json_decode($response);
-        return $decodeResponse->data ?? $decodeResponse;
-    }
-
-    /**
-     * @param $queryParams
-     * @param $formsParams
-     * @param $headers
-     * @return void
-     */
-    public function resolveAuthorization(&$queryParams, &$formsParams, &$headers): void
-    {
-        $accessToken = $this->resolveAccessToken();
-
-        $headers['Authorization'] = $accessToken;
-    }
-
-    /**
-     * @return string
-     */
-    public function resolveAccessToken(): string
-    {
-        return 'Bearer '. $_ENV('BASE_TOKEN');
-    }
 }
